SAM2/requirements.txt
torch
torchvision
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0
matplotlib>=3.9.1
opencv-python>=4.7.0